XMB horizontal browsing tweaks

+++ menu/menu_animation.h
@@ -93,7 +93,7 @@ typedef struct menu_animation_ctx_delta
 
 typedef struct menu_animation_ctx_tag
 {
-   int id;
+   uintptr_t id;
 } menu_animation_ctx_tag_t;
 
 typedef struct menu_animation_ctx_subject
@@ -108,7 +108,7 @@ typedef struct menu_animation_ctx_entry
    float target_value;
    float *subject;
    enum menu_animation_easing_type easing_enum;
-   int tag;
+   uintptr_t tag;
    tween_cb cb;
 } menu_animation_ctx_entry_t;
